practice

+++ deletinglog.sh
@@ -13,9 +13,21 @@ file_delete=$(find $source -type f -mtime +14 -name "*.log")
 
     while IFS= read -r line
     do
-        echo "deleting more than 14 days old logs from $source:$line"
-         rm -rf $line
+       op=$($line|wc -l)
+       if [ $op -ne 0]
+       then
+           echo "deleting more than 14 days old logs from $source:$line"
+           rm -rf $line
+        else
+           echo "no files exist"   
+    
+        fi
+       
     
     done <<< $file_delete
+
+      
+    
+
